Separated tab content form Plans.jsx and created new component called "TabContent".
